refactor: move index to src folder
