+++ package.json
@@ -7,9 +7,8 @@
   "scripts": {
     "test": "jest --silent",
     "test:watch": "jest --watch",
-    "start:dev": "nodemon index.js",
-    "start": "node index.js",
-    "start:prod": "node dist/index.js"
+    "start:dev": "nodemon src/index.js",
+    "start": "node src/index.js"
   },
   "repository": {
     "type": "git",
